+++ source/databricks/geh_stream/aggregation_utils/aggregators/adjust_flex_consumption.py
@@ -13,7 +13,7 @@
 # # limitations under the License.
 from geh_stream.codelists import Colname, ResultKeyName, ResolutionDuration, MarketEvaluationPointType
 from geh_stream.shared.data_classes import Metadata
-from geh_stream.aggregation_utils.aggregation_result_creator import create_dataframe_from_aggregation_result_schema
+from geh_stream.aggregation_utils.aggregation_result_formatter import create_dataframe_from_aggregation_result_schema
 from pyspark.sql import DataFrame
 from pyspark.sql.functions import col, when, lit
